Change namespace for geometry objects
