+++ lib/Geometry/CircularString.php
@@ -21,15 +21,17 @@
  * SOFTWARE.
  */
 
-namespace CrEOF\Geo\Obj;
+namespace CrEOF\Geo\Obj\Geometry;
+
+use CrEOF\Geo\Obj\Geometry;
 
 /**
- * Class Polygon
+ * Class CircularString
  *
  * @author  Derek J. Lambert <dlambert@dereklambert.com>
  * @license http://dlambert.mit-license.org MIT
  */
-class Polygon extends Object implements GeometryInterface
+class CircularString extends Geometry
 {
-    const T_TYPE = self::T_POLYGON;
+    const T_TYPE = self::T_CIRCULARSTRING;
 }
