+++ lib/ObjectFactory.php
@@ -41,14 +41,14 @@ class ObjectFactory implements ObjectFactoryInterface
 
     //TODO: add remaining types
     const C_OBJECT             = 'CrEOF\Geo\Obj\Object';
-    const C_POINT              = 'CrEOF\Geo\Obj\Point';
-    const C_LINESTRING         = 'CrEOF\Geo\Obj\LineString';
-    const C_POLYGON            = 'CrEOF\Geo\Obj\Polygon';
-    const C_MULTIPOINT         = 'CrEOF\Geo\Obj\MultiPoint';
-    const C_MULTILINESTRING    = 'CrEOF\Geo\Obj\MultiLineString';
-    const C_MULTIPOLYGON       = 'CrEOF\Geo\Obj\MultiPolygon';
+    const C_POINT              = 'CrEOF\Geo\Obj\Geometry\Point';
+    const C_LINESTRING         = 'CrEOF\Geo\Obj\Geometry\LineString';
+    const C_POLYGON            = 'CrEOF\Geo\Obj\Geometry\Polygon';
+    const C_MULTIPOINT         = 'CrEOF\Geo\Obj\Geometry\MultiPoint';
+    const C_MULTILINESTRING    = 'CrEOF\Geo\Obj\Geometry\MultiLineString';
+    const C_MULTIPOLYGON       = 'CrEOF\Geo\Obj\Geometry\MultiPolygon';
+    const C_CIRCULARSTRING     = 'CrEOF\Geo\Obj\Geometry\CircularString';
     const C_GEOMETRYCOLLECTION = 'CrEOF\Geo\Obj\GeometryCollection';
-    const C_CIRCULARSTRING     = 'CrEOF\Geo\Obj\CircularString';
     const C_FEATURE            = 'CrEOF\Geo\Obj\Feature';
 
     /**
