+++ tests/Geometry/CircularStringTest.php
@@ -21,11 +21,11 @@
  * SOFTWARE.
  */
 
-namespace CrEOF\Geo\Obj\Tests;
+namespace CrEOF\Geo\Obj\Tests\Geometry;
 
 use CrEOF\Geo\Obj\Configuration;
-use CrEOF\Geo\Obj\CircularString;
 use CrEOF\Geo\Obj\Exception\ExceptionInterface;
+use CrEOF\Geo\Obj\Geometry\CircularString;
 use CrEOF\Geo\Obj\Object;
 
 /**
@@ -64,8 +64,8 @@ class CircularStringTest extends \PHPUnit_Framework_TestCase
 
         $circularString = new CircularString($value);
 
-        if (! array_key_exists('value', $expected)) {
-            self::assertEquals($expected, $circularString->getValue());
+        if (! array_key_exists('coordinates', $expected)) {
+            self::assertEquals($expected, $circularString->getCoordinates());
         } else {
             foreach ($expected as $property => $expectedValue) {
                 $function = 'get' . ucfirst($property);
@@ -82,9 +82,9 @@ class CircularStringTest extends \PHPUnit_Framework_TestCase
     {
         return [
             'testGoodWkbCircularString' => [
-                'value'      => pack('H*', '01080000000300000000000000000000000000000000000000000000000000f03f000000000000f03f00000000000000400000000000000000'),
-                'validators' => null,
-                'expected'   => [[0,0],[1,1],[2,0]]
+                'coordinates' => pack('H*', '01080000000300000000000000000000000000000000000000000000000000f03f000000000000f03f00000000000000400000000000000000'),
+                'validators'  => null,
+                'expected'    => [[0,0],[1,1],[2,0]]
             ],
         ];
     }
